Refactor tests with new file paths
