+++ test/client/App.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { shallow } from 'enzyme';
-import App from '../../../client/src/App.jsx';
+import App from '../../client/src/App.jsx';
 
 test('<App />', () => {
