--- test/client/containers/SearchResults.test.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { shallow } from 'enzyme';
-import SearchResults from '../../../client/src/containers/SearchResults.jsx';
-
-test('<SearchResults />', () => {
-  it('renders SearchResultsItems', () => {
-    const SearchResults = shallow(<SearchResults />);
-    expect(SearchResults.find('SearchResultsItem')).toBe(true);
-  });
-});
